requirements.txt
# Web Framework
fastapi==0.117.1
uvicorn==0.37.0

# Database
SQLAlchemy==2.0.43
asyncpg

# Authentication & Security
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv

# Data Validation
pydantic==2.11.9

# WebSocket & Real-time
python-socketio==5.13.0
python-engineio==4.12.2

# Rate Limiting
slowapi

# Form Data (required for OAuth2PasswordRequestForm)
python-multipart==0.0.20
